requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.7.0.72
python-dotenv==1.0.0
drf-yasg==1.21.5
django-filter==23.1
djangorestframework-simplejwt==5.2.2 